+++ data/examples/declaration/value/function/arrow/proc-lambdas-out.hs
@@ -5,5 +5,6 @@ foo = proc a -> \f b -> a -< f b -- Foo
 bar =
   proc x -> \f g h ->
     \() ->
-      \(Left (x, y)) -> -- Tuple value
+      \(Left (x, y)) ->
+        -- Tuple value
         f (g (h x)) -< y