+++ data/examples/other/no-linear-arrows-out.hs
@@ -1,5 +1,0 @@
-type a % b = (a, b)
-
-type Foo a m b = a % m -> b
-
-type Bar a m b = a % m -> b